@@ -99,48 +99,48 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~13.2.2",
-    "@angular/common": "~13.2.2",
-    "@angular/compiler": "~13.2.2",
-    "@angular/core": "~13.2.2",
-    "@angular/forms": "~13.2.2",
-    "@angular/platform-browser": "~13.2.2",
-    "@angular/platform-browser-dynamic": "~13.2.2",
-    "@angular/router": "~13.2.2",
-    "@fortawesome/angular-fontawesome": "^0.10.1",
-    "@fortawesome/fontawesome-svg-core": "~1.3.0",
-    "@fortawesome/free-solid-svg-icons": "^6.0.0",
-    "angular-split": "^13.2.0",
-    "rxjs": "~7.5.4",
+    "@angular/animations": "^18.2.0",
+    "@angular/common": "^18.2.0",
+    "@angular/compiler": "^18.2.0",
+    "@angular/core": "^18.2.0",
+    "@angular/forms": "^18.2.0",
+    "@angular/platform-browser": "^18.2.0",
+    "@angular/platform-browser-dynamic": "^18.2.0",
+    "@angular/router": "^18.2.0",
+    "@fortawesome/angular-fontawesome": "^0.15.0",
+    "@fortawesome/fontawesome-svg-core": "~6.6.0",
+    "@fortawesome/free-solid-svg-icons": "^6.6.0",
+    "angular-split": "^17.2.0",
+    "rxjs": "~7.8.1",
     "semver": "^7.3.5",
     "tslib": "^2.3.0",
-    "xml2js": "^0.4.23",
-    "zone.js": "~0.11.4"
+    "xml2js": "^0.6.2",
+    "zone.js": "~0.14.10"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^13.2.3",
-    "@angular-eslint/builder": "13.1.0",
-    "@angular-eslint/eslint-plugin": "13.1.0",
-    "@angular-eslint/eslint-plugin-template": "13.1.0",
-    "@angular-eslint/schematics": "13.1.0",
-    "@angular-eslint/template-parser": "13.1.0",
-    "@angular/cli": "~13.2.3",
-    "@angular/compiler-cli": "~13.2.2",
-    "@types/jest": "^27.4.0",
-    "@types/node": "^17.0.18",
+    "@angular-eslint/builder": "18.3.0",
+    "@angular-eslint/eslint-plugin": "18.3.0",
+    "@angular-eslint/eslint-plugin-template": "18.3.0",
+    "@angular-eslint/schematics": "18.3.0",
+    "@angular-eslint/template-parser": "18.3.0",
+    "@angular/build": "^18.2.0",
+    "@angular/cli": "^18.2.0",
+    "@angular/compiler-cli": "^18.2.0",
+    "@types/jest": "^29.5.12",
+    "@types/node": "^22.3.0",
     "@types/semver": "^7.3.9",
     "@types/vscode": "^1.64.0",
-    "@types/xml2js": "^0.4.9",
-    "@typescript-eslint/eslint-plugin": "5.12.0",
-    "@typescript-eslint/parser": "5.12.0",
-    "eslint": "^8.2.0",
-    "jest": "^27.5.1",
-    "jest-preset-angular": "^11.1.1",
-    "npm-watch": "^0.11.0",
-    "ts-loader": "^9.2.6",
-    "typescript": "~4.5.2",
+    "@types/xml2js": "^0.4.14",
+    "@typescript-eslint/eslint-plugin": "^8.1.0",
+    "@typescript-eslint/parser": "^8.1.0",
+    "eslint": "^9.9.0",
+    "jest": "^29.7.0",
+    "jest-preset-angular": "^14.2.2",
+    "npm-watch": "^0.13.0",
+    "ts-loader": "^9.5.1",
+    "typescript": "~5.5.4",
     "webpack": "^5.69.1",
-    "webpack-cli": "^4.9.2"
+    "webpack-cli": "^5.1.4"
   },
   "jest": {
     "preset": "jest-preset-angular",
